+++ HandHistories.SimpleObjects/Entities/PlayerHistory.cs
@@ -12,7 +12,7 @@ namespace HandHistories.SimpleObjects.Entities
         public int Id { get; set; }
         //FK:
         [Required]
-        public int GameNumber { get; set; }
+        public ulong GameNumber { get; set; }
         [Required]
         public string PlayerName { get; set; }
         [Required]
